tuningfork: List initializers in execution order
This commit does not change current behavior.

'env_' will be initialized last, since it's the last of this
class's three members.  We change the ordering in which we write
our initializers to reflect this, and as a heads up to any future
work that might try to initialize 'mgr_' or 'jmgr_' based on 'env'.
This also gets rid of the Wreorder warning some compilers would
give on this code.

Test: TreeHugger
Change-Id: I3d8ee7990c4f6a7e38b0068e9bb189c152e6d182