SwappyCommon: Initialize fields in order
We move the initialization of 'mRefreshPeriod' into its actual
spot in terms of when it gets initialized.

This doesn't change current behavior, and there's no current bug
since 'mChoreographerFilter' uses 'refreshPeriod' in its
initialization.  We just make this change to help future-proof
this code, and be clearer about initialization ordering.

Test: TreeHugger
Change-Id: Icf53406dc8aac07584a064ff088279a7019ed252